Assert on stack height diff instead of total stack height
With this commit, the assertion on stack height is imposed on the
difference in the stack height that the code generates instead of the
total (final) stack height. This should make composition of code easier
and the `execute` function more flexible as it asserts less on the
initial state on the stack when the code execution begins.