Update README with new cycle counts after fixing jump stack problems
We shouldn't normally overwrite the lines in the README file but since
the previous implementations did not leave the jump stack unchanged,
they are now considered buggy. And there's no reason to have wrong
algorithms/subroutines listed in the README. Therefore, I think it's
justified and preferable to update these entries.