Fix discrepancy in memory manipulation by rust-shadowing of MMR function
A memory discrepancy is fixed by rewriting Rust-shadowed version of the
`calculate_new_peaks_from_append` function to better match what's going
on in the TASM code. Prior to this commit, the TASM code had some junk
in memory that the Rust-shadowed version did not have. This was because
of how `pop` was implemented in TASM. `pop` is in TASM written such
that only the length indicator was decremented. The old element isn't
actually removed. With this commit, the Rust-shadowing reflects this
behavior.

This commit also adds helper functions for list manipulation in the
Rust-shadowed functions.

Co-authored-by: Simon Shine <simon@neptune.cash>