switch input argument order on MMR functions
This was done to help me copy (dup) a lot in the
verify_from_secret_input_through_memory function. I'm not sure that
those kinds of optimizations are so clever though. They only save up to
ten instructions, and take a lot of time and have a dubious effect.

Going forward we should focus on getting the code to work before we
attempt to bring the cycle count down with these small tweaks.