Add type annotations to all snippets
This got a bit complicated because of the need for list supports. Since
we want to know what the type of an element in a list is, I added data
type to the list snippet structs. But then the interface for some of the
functions changed, since `&self` was now needed for some of the
functions. This lead to a rewrite of all the tests, since they now
needed `&self: &Snippet` as input.

I still think this is a good idea though, as we now from external
dependencies can see if a function signature changes.