Use right-shifting snippet in popcount snippet
The code *seemed* to work prior to this commit, but it didn't follow the
algorithm found on Wikipedia and I don't really trust it when it
doesn't. This increases the cycle count for the popcount snippet by 10 %
but I think the higher trust that it works makes it worth it.