Improve snippet for adding two u64s
We can assume that the inputs are valid u64s, so the carry is either a
zero or a one, and this value can just be added to the higher part of
the input directly. There's no need for conditional code.

This decreases table heights for many snippets, and increases it for
none.