Make MMR benchmark sizes as uniform as possible
Common case is defined to be an MMR with 2^31 leafs. Worst case is
defined to be an MMR with 2^62 leafs.

This change should make it fairly easy to spot how the runtime scales
with the size of the MMR: If worst case is ~twice as big as common case,
there is no significant constant factor and the runtime scales with the
height of the MMR, which is the logarithm of the leaf count.