Add assert so  that u64 div-mod fails in 0-divsion for large numerators
Previously the execution only failed with 0-division if the numerator
was below 1^32.

With this change, I believe u64 div-mod is correct. Although it is very
inefficient. Any practical application would prabably divine in the
result and prove that it is true instead of this code that uses 10'000
clock cycles in the worst case.