Delete benchmarker's own VM execution wrapper
The benchmarker had its own execution wrapper which did not take static
memory allocation into account. This meant that some of the benchmark
(notably the MMR append algorithm's) crashed due to a stack underflow
since the tests and the benchmark did not agree on a specific list
pointer. Instead of fixing the benchmarker's VM wrapper, I opted to only
have one way of compiling assembly, importing libraries, and
interacting with the VM.

This might lead to a off-by-one error/update in benchmarks, since the
benchmarks were (correctly) calculating the height of the produced
tables, whereas the test wrapper was keeping track of the cycle count,
which is the processor table's height minus one.