Make initializing dyn malloc by `execute` optional
External callers (like the `tasm-lang` compiler) should probably output
TASM that initializes the dynamic allocator, whereas the execution of
snippets in this repository do not need to do that, as the snippets are
tested in isolation and not in a wider context. For this reason, we
allow the caller to decide what `execute` should do.