Fix off-by-one error in initialization on dyn malloc
The initial value of dyn malloc must be set to whatever value `kmalloc`,
the static allocator would have returned if it was called at that time.

Prior to this commit, the dyn malloc was initialized to the number of
words allocated by the static allocator, which was 1 too low a number
which lead the two types of memory to interfere.