Make snippet-export check optional
We have some snippets we don't necessarily need to export for use in the
compiler or in other dynamic code. So we move the check that the
snippets being tested is exported through "all_snippets" and make it
optional, with a boolean parameter to select for this check.

This also means we can remove the export of some test-snippets.