Generalize `map` to be able to handle more arguments
With this commit, the higher-order function `map` can now take more
input arguments than just the list element.

Extreme care has to be taken when defining the inner function though, as
there is additional elements on the stack that need to be invariant
under the inner function, *and* the additional arguments need to be dug
out from three words down the stack. See the added test cases for
working examples.