Add new InnerFunctionType `NoFunctionBody`
This new enum variant should be used when the inner function is defined
somewhere else, when it's not the `Map` snippets responsibility to
import the inner function.

This is relevant if you e.g. define a local function somewhere that you
want to use in a `map` and potentially also from some other part of the
code.