Make program generation deterministic
All imports were collected in `HashMap`/`HashSet` and when they were
looped over to generate the code for all imports, the imported snippets
would be produced in an arbitrary order, resulting in different
programs/program hashes. To fix this, we sort the imported snippets
alphabetically before we produce all the dependency code.