Change interface for verifying that VM execution crashes
Previously we verified that VM execution crashes by decorating tests
with `#[should_panic]` attributes, but the called test functions could
panic for many reasons, not just VM execution failure. With this commit,
we now get an error if and only if VM execution fails, if anything else
goes wrong, the called function panics. In the tests, we can then remove
the clumsy `should_panic` attribute and replace it with a
`assert!(vm_execution_res.is_err())`. It's still not perfect, but should
be a lot more robust than what we had before.