add and derive `decode_iter` on trait `TasmObject`
The new method `decode_iter` allows you to build an object of any type
that implements `TasmObject` from an `Iterator` over `BFieldElement`s.
Previously, `BFieldCodec::decode` would only accept a slice of
`BFieldElement`s, which is difficult to use if you don't know the
length of the slice to feed to it.