Remove `Display` implementation of `DataType`, add `variant_name`
For the `tasm-lang` compiler to be able to generate snippets, it needs
to output a string representing the data types used in `tasm-lib`. This
adds such a method.

Also: The `Display` implementation is removed. It was only used for
entrypoint names, and there we need to use `label_friendly_name` anyway
in order to avoid creating labels that Triton VM's assembler rejects.
Also: `Display` on this type is ambiguous, as there are multiple
ways to express a DataType value as a string.