fri_verify: verify A leafs
This commit also fixes a bug related to the static allocation
resulting from `xfe_ntt` spilling two words. It was spilling them
into the memory designated for dynamic allocation because the linker
was initialized with the number of words allocated to represent the
memory initial state. The linker is now invoked with the correct
argument (0).