Fix memory initialization in deprecated benchmarks
Memory now has to be initialized through non-determinism before the VM
is run. We no longer have the state-preparation code to put the VM into
the expected start-state.

We fix this by merging memory and non-determinism and verifying that only one
of them is set. But ideally we should entirely remove `memory` from the
`ExecutionState` data structure which the trait functions for snippet
implementations populate.