Remove parameter for statically allocated words from Rust shadowing
This parameter was unused. In the general case, the information about
the number of statically allocated words is *not* present when the
Rust-shadowing is invoked, as this value is only known internally by the
TASM-linker once all code snippets have been assembled. Since the
information is not available at the time, the Rust-shadowing logic
cannot depend on it.