Prevent use of higher-order functions with more than one input/output type
@aszepieniec and I were fixing a bug in neptune-core when I changed
the function signature of a snippet used as an inner function for `map`
from one taking a tuple as input to one taking two input arguments.
Instead of getting a compile-time error when the assembly code was
stringed together, the code did something wrong. Hopefully these
`assert` statements that I add in this commit will prevent other people
from wasting 20 minutes on this like we just did.