Add implementation of `SpongeHasher::init` function
This mainly serves to make it easy for the `tasm-lang` compiler to call
all functions from the `SpongeHasher` trait. The implementation is trivial,
and any function call can simply be inlined if the developer is so
inclined. Inlining the `init` function would mean that there is *no*
overhead associated with defining this as a snippet.