fix: Bug in test-setup for list's `get` snippet
In the test setup list length and index was calculated consistently, but the
list length was not passed to the function to setup the test-case. Instead a
new random value was passed to this function resulting in the possibility
of an out-of-bounds error, as the index could be greated than or equal to the
list length used in the state-preparation function.