feat: Add snippet for squeezing repeatedly for statically known squeeze count
This new snippet halves the clock cycle count compared the the snippet
where the squeeze count is only known at run-time. But that improvement
comes at a very high hash-table cost, as the hash table for this
isolated snippet is twice the clock cycle count.

The snippet is implemented by unrolling the inner loop, so there's quite
a lot of code.