docs: Improve error message on memory-equivalence failure
Similar to 228375fe49ab215ccc0cffd50aed7d1621d1f342, we would like to
see where the memory in the error messages are coming from. So instead
of calling memory representations `A` and `B`, we now have to name the
representations. The names will often be "Rust-shadow" and "TVM
execution".