refactor: Change memory convention for `Challenges`
Store `Challenges` in the 3-rd to last memory page, which is the 2nd-to
-last *full* memory page of size 2^32.  As a reminder, the state of
the dynamic allocator lives in the last memory page, the memory used
for the AIR constraint valuation calculation lives in the 2nd-to-last
page, and `Challenges` now lives in the 3-rd to last page. The address
for this page is `((1 << 32) - 3) * (1 << 32)`.