feat: Add `split_off` snippet
This snippet is used by the current version of the verifier. So it's
convenient to have it as a snippet as well.

Technically, the verifier uses the slice method `split_at`, but we
`split_off` can be used for a similar thing for lists, and we don't
currently support a concept of slices, so `split_off` will have to work
as a replacement.