docs: Add more table heights to benchmark results
Previously we only recorded clock cycle count, hash-table height, and
u32-table height. But when attempting to run a recursive
proof-verification through `tasm-lang`, we discovered that the op-stack
table height actually dominates, and not the processor table height
(clock cycle count). Therefore, we need to also record the op-stack
table height, and we're recording the RAM table height as well now that
we're at it.

This new data should function as a starting point for a new round of
optimizations, one targeting the opstack-table height!

We need that for fast recursion!