feat: HABEMUS (2nd) VERIFIER!!
Fix remaining part of the Triton-VM proof verifier. This is our 2nd
version of the verifier, the 1st being developed in
<https://github.com/TritonVM/tasm-lang>. This version is written
entirely in assembler, forcing us to think about every single instruction,
clock cycle, and opstack table row. I imagine that the version of the
verifier that we'll use in production will come from the compiler but
will make heavy use of snippets from `tasm-lib` to bring down the table
heights. This verifier ensures that all such snippets are available to
the developer.

This version of the verifier seems to use around 1/6 fewer clock cycles
than the original compiler-developed version.