feat: Disallow very big proof items
To prevent a potential wrap-around in the memory adddresses or other
weird stuff like reading from dynamically-initialized memory in
relation to the proof-iter and reading of a proof from
non-deterministic memory, a check is added that the indicated size
of a proof item does not exceed a certain threshold. This threshold
was somehwat arbitrarily chosen to be $2^22$, as I think that's
plenty.