test: Add basic test of and rename `VerifyTableRow`
Snippet is renamed to reflect the fact that it is used for base rows,
extension rows, and quotient segments.

A very basic (but complicated to write) test is added to increase our
assurance that this snippet does what we expect it to do, but we ought
to add negative tests as well to verify that the snippet crashes the VM
if it is fed a bad authentication path or a leaf index that's not a u32.