fix: Adapt STARK verifier to Triton VM prover v0.40
A slight change was made in Triton VM's STARK protocol, as the
base/ext/quotient *and* deep codeword weights are now all sampled
through one sponge operation, whereas this was previous two, one for
base/ext/quotient and another one for the deep codeword weights.

This saves 18(!) in the dominating table height, which is the op-stack
table, so it's not exactly a huge performance boost but admittedly a bit
cleaner way of sampling weights.