perf: Improve stack-management when verifying FRI (xfe,idx) auth paths
Previously, we were dupping the root to the top of the stack, and then
throwing it away after a call to `assert_vector`. Now we don't throw it
away.

Reduces the cost you pay for each addition round of FRI by ~12 %.

Previous (inner-padded-height, opstack-table-height):
256  => 392969
512  => 414331
1024 => NA
2048 => 459281
Cost of three rounds: 66312

With these changes:
256  => 382211
512  => 400883
1024 => 420297
2048 => 440458
Cost of three rounds: 58247

Savings: 12.16 %.

We could save a few more percentages by verifying the list from highest
index to lowest but I deemed that too big a change, as it might even
affect upstream Triton VM.