perf: use new instruction `recurse_or_return`
This new instruction significantly reduces the rows that are added due
to Merkle authentication path verification. The reference STARK verifier
with a FRI expansion factor of 4 and an inner padded height of $2^20$
sees a 19.0 % reduction in its current bottleneck table, the processor
table.

Inner padded height: 2^{20}
FRI EF: 4
Table heights prior to this commit
Processor:       559500
Opstack:         476200
RAM:             309131
Hash:            249025 (total)
U32:             123527

Table heights after this commit
Processor:       453023
Opstack:         370696
RAM:             310696
Hash:            249943 (total)
U32:             124686

With this commit, FRI-verification is responsible for 70.5 % of the
table rows in the processor table, which is the bottleneck. Let's
attempt to optimize the FRI verifier next.