perf: Remove deprecated copying of last codeword in FRI
The last codeword was copied because we used do a NTT transformation on
it, which would mutate the list. Now that we've replaced NTT with
barycentric evaluation, we don't need to copy the list, as barycentric
evaluation does not mutate its input.

This saves around 8000 rows in the provessor table and the same in the
op-stack table for EF=4. More for higher expansion factors.