perf: reduce memory allocation used in `MerkleRoot` snippet
Reduces memory allocation when calculating Merkle roots.

`MerkleRoot` now only allocates from memory once instead of one time for
each layer in the Merkle tree. This reduces the processor row count
minimally but also results in a cleaner execution, as the `DynMalloc`
snippet is invoked a maximum of two times when finding a Merkle root
from a list of X-field elements.