WIP: Eliminate most higher-order calls in FRI
All higher-order functions calls used within each round of FRI is with
this commit eliminated. In the current version the authentication paths
of the B-values are not verified, so this FRI is currently not sound.

A few higher-order function calls still happen but none within each FRI
round. This FRI-verifier is going to get fast!