Rewrite barycentric to use more RAM rows, fewer processor rows
Rewrite the snippet for barycentric evaluation to use fewer processor
and opstack rows, and more RAM rows. The goal with this rewrite is to
make higher expansion factors more feasible, as the very high row count
for this snippet made expansion factors above 8 meaningless. With this
rewrite, expansion factors above 8 can be considered.

This rewrite is not completely without controversy, as it increases the
RAM table row-count somewhat. For expansion factor 4, RAM table row
count is increases by 4'600 and twice that for expansion factor 16. The
savings to the processor table are 7'000 and 20'000, respectively.

If these changes are not beneficial, this commit can be reverted.