perf: Use `recurse_or_return` in inner-most loop of `MerkleRoot` snippet
Save another 11 % of rows in the processor table for the relevant
problem size. This is close to being the end-of-the-line for efficient
Merkle root calculation, I believe. Even fewer lines could be used if
the leaf number is known, by using the snippet for statically-known leaf
counts, but you pay a big price for program size (and thus hash table
row count) if you do that.

This completes the 1st optimization mentioned in #106.