test(contains): Verify we got parity right
Adds a test that explicitly checks that the snippet returns false when
testing for membership of a mirrored value. This test ensures that we
didn't accidently flipped parity in both the TASM code *and* in the Rust
shadowing, as a double bug there would cancel each other thus resulting
in false positive tests for all test runs.