perf!: Faster and const time u64::lt snippet
Improve worst-case performance of `lt` snippet for `u64`. This snippet exists
in two versions, one that consumes its input arguments, and one that does not.

Both snippet's running time is improved with this change for the worst-case
inputs but the snippet that preserves the arguments goes from 12 to 14 clock
cycles in the common (best) case scenario. The arg-consuming snippet is
improved for both the best-case and worst-case scenario.