mmr::authentication_struct: implement snipet and disallow empty list of leafs
When calculating the root, don't allow the list of leafs to be empty.
We think we never need the snippet with those inputs, so that should be
fine. It makes the initialization of the `t` value (that should end up
containing the root) simpler, as we can do:

```rust
let mut t = indexed_leafs.first().unwrap().1;
```

instead of
```rust
        let mut t = auth_struct
            .first()
            .copied()
            .unwrap_or_else(|| indexed_leafs.first().unwrap().1);
```

Code works with sane test.