feat(stark_verify): Allow dynamic memory layouts for STARK verify
This makes it possible to verify multiply proofs, i.e. to make a program
that calls the STARK verifier multiple times.

In the previous version, it was assumed that the position in RAM of the
out-of domain rows was known at compile. But that doesn't work if you're
verifying a dynamic number of proofs.

The extra cost of verifying a proof with a dynamic memory layout comes
from the AIR evaluation code. The cost is independent of padded height
and is around 10k rows. See the benchmarks in this commit for more
precise row counts.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>