fix: Drop population of memory from `update_nondeterminism`
The function previously stored the STARK proof in nondeterministically
initialized memory; now it no longer does. It only appends to the
stream of nondeterministic digests. The doc comment reflects this.

Also, the test testing two proofs uses field getters into a custom
struct to access the proofs, relfecting how the snippet is used in
Neptune-core.

Co-authored-by: Thorkil Værge <thor@neptune.cash>