feat(tasm_object): Add bounds-checks to auto-derived field getters
Since `TasmObject` is defined in this repo, and `MmrAccumulator` et al
are defined upstream, we need to auto-generate some code to implement
this trait. This commit updates the auto-generated code.

This bounds check only adds eight rows to the processor table for the
verifier. More for mmr successor proof since it uses these field getters
extensively and repeatedly.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>