refactor(TasmObject): Reduce max dyn-field size indicator by factor 4
New max-size: 2^28
Back-of-the-envelope calculations:
each proof is ~120.000 words, so if you want to verify 1.000 proofs in
one execution, that would result in an witness size of 1.2*10^8 which is
less than 2^28 ~= 2.5 * 10^8.

Verifying 1.000 proofs, with the current performance of the verifier,
would result in a max table height of 2.5 * 10^8 for FRI expansion
factor 4, which gives a FRI domain length of 2*10^9. Since 2^31 ~=
2*10^9 is the highest FRI domain length we currently support (because
the FRI domain length needs to be a valid u32 and a power of 2), this
reduction in the max size seems reasonable to me.

This max size means that a witness initialized at address 0 would need
to contain 16 dynamically-sized fields to escape into the next memory
page. the highest number of fields, we are currently using is
[`ProofCollection`](https://github.com/Neptune-Crypto/neptune-core/blob/asz/transaction-consensus/src/models/blockchain/transaction/validity/proof_collection.rs)
which has eight dynamically-sized fields.