feat(traits): Add new trait `mem_preserver`
This trait makes it easier to reason about code, as there is a snippet
(verify_mmr_successor) which can read from ND but does not modify
memory. In order to reason about the soundness of that snippet, this
trait `MemPreserver` is added.

It makes it easier to clarify that the field-access of
`verify_mmr_successor` is safe.