chore: update benchmarks for TVM0.42-alpha10 -> 0.42 upgrade
The relevant difference in performance is caught by the diff in the
Triton VM repo where the instructions `pick` and `place` are introduced:
https://github.com/TritonVM/triton-vm/commit/b76939226968453c8966d327034f8b343390867c

The main cost is captured by the STARK-verify benchmark, showing a 2.813
increase in the processor table row count and a 3.623 increase in the
RAM table. These increases are less than 2 % of the total row count, so
these new instructions seem worthwhile! Very nice!