Handle new fields in VM proof item
The fields introduced in the previous commit,
"refactor: Add more metadata to vm_proof_iter struct" are with this
commit updated/handled correctly. This means that the
`current_item_count` is incremented once for every time a proof item is
dequeued. The purpose is to compare the counted value with the reported
value at the end of verification.