refactor: Update generated Tasm-object derivations
Also, improve maintainability by using a different approach.

Previously, the type in question was re-implemented manually, then
`cargo expand` was used to derive the required code, which was then
copy-pasted into the repository.

Now, dummy structs are used to derive the code, and all method calls are
forwarded to those dummy structs. This ensures that the derived code is
always up to date. Differences in the actual type and the
re-implementation are caught at compile time.

fix #110