feat(TasmStruct): Add destructuring capabilities
When writing Triton assembly involving `TasmStruct`s, it is common to
access many if not all fields of of the struct. Previously, the only
sensible way to do so was duplicating the struct pointer, then
executing the field-getter.

Now, executing the code emitted by the new `destructure` function
replaces a struct pointer with pointers to all its fields.